requirements.txt
SQLAlchemy~=1.4.39
PyQt5~=5.15.7
Sphinx
pytest~=7.1.2
pytest-cov
pytest-pep8
pytest-timeout
pytest-aiohttp
numpy
pycodestyle
Pillow~=9.2.0
aiohttp~=3.8.1
cchardet
aiodns
aiohttp_security
motor~=3.0.0
aiohttp_session[secure]
cryptography~=37.0.4
bson~=0.5.10
quamash
sqlalchemy_aio
Django~=4.0.6
pymongo~=4.2.0
setuptools==63.3.0